+++ plugins/wiggle/src/CreateMultiWiggleExtension/ConfirmDialog.tsx
@@ -1,13 +1,12 @@
 import React, { useState } from 'react'
 import {
   Button,
-  Dialog,
   DialogActions,
   DialogContent,
-  DialogTitle,
   TextField,
   Typography,
 } from '@mui/material'
+import Dialog from '@jbrowse/core/ui/Dialog'
 import {
   AnyConfigurationModel,
   readConfObject,
@@ -23,8 +22,11 @@ const ConfirmDialog = ({
   const [val, setVal] = useState('MultiWiggle ' + Date.now())
   const allQuant = tracks.every(t => t.type === 'QuantitativeTrack')
   return (
-    <Dialog open onClose={() => onClose(false)}>
-      <DialogTitle>Confirm multi-wiggle track create</DialogTitle>
+    <Dialog
+      open
+      onClose={() => onClose(false)}
+      title="Confirm multi-wiggle track create"
+    >
       <DialogContent>
         <Typography>
           {!allQuant
